feat: Add Hono backend with IP logging and caching
- Implemented a Hono server (`server/index.ts`) running on port 3001.
- Setup SQLite database (`gittalent.db`) using `better-sqlite3` to store:
  - Analysis logs (IP address, username, timestamp).
  - Cached profiles (username, data, timestamp).
- Moved analysis logic from frontend to backend (`server/lib/`).
- Updated frontend `services/analyzer.ts` to fetch from `/api/analyze`.
- Configured Vite proxy to forward `/api` requests to the backend.
- Added `npm run server` script.
- Secured `DEEPSEEK_API_KEY` by moving it to the backend.
- Implemented 7-day cache validity for analyzed profiles.
