+++ .gitignore
@@ -28,3 +28,6 @@ TESTING_GUIDE.md
 IMPROVEMENTS.md
 ANALYSIS_SUMMARY.md
 ARCHITECTURE.md
+
+# Database
+*.db
