+++ vite.config.ts
@@ -8,6 +8,12 @@ export default defineConfig(({ mode }) => {
       server: {
         port: process.env.PORT ? parseInt(process.env.PORT) : 3000,
         host: '0.0.0.0',
+        proxy: {
+          '/api': {
+            target: 'http://localhost:3001',
+            changeOrigin: true,
+          }
+        }
       },
       preview: {
         port: process.env.PORT ? parseInt(process.env.PORT) : 3000,
